+++ Packages/EventsWidgetLogic/Sources/EventsWidgetLogic/EventsTimeline.swift
@@ -1,0 +1,15 @@
+import Foundation.NSDate
+
+public struct EventsTimeline: Equatable {
+    
+    public var entries: [EventTimelineEntry]
+    
+    public var snapshot: EventTimelineEntry {
+        entries.first ?? EventTimelineEntry(date: Date(), events: [], additionalEventsCount: 0)
+    }
+    
+    public init(entries: [EventTimelineEntry]) {
+        self.entries = entries
+    }
+    
+}
