+++ Packages/EventsWidgetLogic/Sources/EventsWidgetLogic/EventsTimelineController.swift
@@ -14,20 +14,6 @@ public struct EventsTimelineController {
 
 // MARK: - Timeline
 
-public struct EventsTimeline: Equatable {
-    
-    public var entries: [EventTimelineEntry]
-    
-    public var snapshot: EventTimelineEntry {
-        entries.first ?? EventTimelineEntry(date: Date(), events: [], additionalEventsCount: 0)
-    }
-    
-    public init(entries: [EventTimelineEntry]) {
-        self.entries = entries
-    }
-    
-}
-
 extension EventsTimelineController {
     
     public struct TimelineOptions {
